Refactor/reformat codebase and fix cirrus warnings

+++ app/src/androidTest/java/sdp/moneyrun/map/CoinGenerationHelperTest.java
@@ -15,8 +15,8 @@ import static org.junit.Assert.assertEquals;
 public class CoinGenerationHelperTest {
 
     @BeforeClass
-    public static void setPersistence(){
-        if(!MainActivity.calledAlready){
+    public static void setPersistence() {
+        if (!MainActivity.calledAlready) {
             FirebaseDatabase.getInstance().setPersistenceEnabled(true);
             MainActivity.calledAlready = true;
         }
@@ -38,11 +38,11 @@ public class CoinGenerationHelperTest {
         loc.setLongitude(4);
         loc.setLatitude(8);
         int radius = 1000;
-       for(int i = 0; i < 100000; i++){
-           Location random = CoinGenerationHelper.getRandomLocation(loc, radius);
-           double distance = MapActivity.distance(loc.getLatitude(), loc.getLongitude(), random.getLatitude(), random.getLongitude());
-           assert(distance < radius && distance > MapActivity.THRESHOLD_DISTANCE);
-       }
+        for (int i = 0; i < 100000; i++) {
+            Location random = CoinGenerationHelper.getRandomLocation(loc, radius);
+            double distance = MapActivity.distance(loc.getLatitude(), loc.getLongitude(), random.getLatitude(), random.getLongitude());
+            assert (distance < radius && distance > MapActivity.THRESHOLD_DISTANCE);
+        }
 
 
     }
@@ -61,16 +61,17 @@ public class CoinGenerationHelperTest {
         Location coinLoc = new Location("");
         centerLoc.setLatitude(1);
         centerLoc.setLongitude(1);
-        double distance = (double) MapActivity.distance(coinLoc.getLatitude(),coinLoc.getLongitude(),centerLoc.getLatitude(),centerLoc.getLongitude());
-        int value = CoinGenerationHelper.coinValue(coinLoc,centerLoc);
-        assertEquals(Math.ceil((distance)/CoinGenerationHelper.VALUE_RADIUS),value,0);
+        double distance = MapActivity.distance(coinLoc.getLatitude(), coinLoc.getLongitude(), centerLoc.getLatitude(), centerLoc.getLongitude());
+        int value = CoinGenerationHelper.coinValue(coinLoc, centerLoc);
+        assertEquals(Math.ceil((distance) / CoinGenerationHelper.VALUE_RADIUS), value, 0);
     }
+
     @Test(expected = NullPointerException.class)
     public void coinValuesFailsOnNullCoinLoc() {
         Location centerLoc = new Location("");
         centerLoc.setLatitude(0);
         centerLoc.setLongitude(0);
-        int value = CoinGenerationHelper.coinValue(null,centerLoc);
+        int value = CoinGenerationHelper.coinValue(null, centerLoc);
 
     }
 
@@ -79,7 +80,7 @@ public class CoinGenerationHelperTest {
         Location coinLoc = new Location("");
         coinLoc.setLatitude(0);
         coinLoc.setLongitude(0);
-        int value = CoinGenerationHelper.coinValue(coinLoc,null);
+        int value = CoinGenerationHelper.coinValue(coinLoc, null);
 
     }
